Make logins less touchy to cookie changes.
First off, the cookie list internal to getCookieString is now a dict instead of a list, this has slightly worse code for formatting, but rather better code for checking to see if we have logged in successfully.

Second, the login declined case no longer tries to format more variables that are provided, allowing us to go down the path correctly.

This gets wikipediafs working again for me against mediawiki 1.19.14+dfsg-1.